Allow slaves to substantiate in parallel
BuildRequestDistributor's _maybeStartBuildsOnBuilder() method is
decorated with @defer.inlineCallbacks.  This means that when it
uses "yield bldr.maybeStartBuild", the entire slave substantiation
process is sequentialized at that point.

To enable parallel slave substantiation, we need to move the error
handling code into a callback function that can be added to the
Deferred that is returned by bldr.maybeStartBuild().

Fixes #4

Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>