Customizable FileUpload step description and url text.
When specifying a url for the FileUpload build step,
the link text would be the filename provided as part
of the masterdest path. In some cases, this is
undesirable. Instead, allow for a url text to be
provided which will be used in place of the
filename in the waterfall view. Also, allow for
the description to be customizable as well.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>