Allow independant EC2 price_multiplier or max_spot_price usage
Change the EC2LatentBuildSlave to allow the use of just one
of either max_spot_price or price_multiplier.  In other words,
either value can be set to None so that only the other value
is used.  When spot instances are used, at least one of those two
parameters must be set.

If only max_spot_price is set, that is the value that is used
as the bid price.  There is not need to look up the historical
bid prices in that case.  It seems likely that in the real world
most people will want to set the maximum price they are willing
to pay.  Artificially bidding lower than one is willing to pay
based on historical information is probably not a very common use
case.

Also, change the logic to use as the bid price the minumum of either
the multiple of the historical average or the max_spot_price.  The
previous behavior was to raise an exception if the multiple was in
excess of the max_spot_price.  That doesn't make much sense.  The
historical average tells us little about the price _now_, so it probably
never makes sense to abort early without at least making an attempt
to substantiate with the max_spot_price.

Fixes #2898 problems 1 and 3.