Handle missing instances during termination
Added exception handling in EC2 slave termination to handle
when an instance is not found. It is possible for instances
to be lost (ex. spot instances) which can cause an EC2 slave
to fail during termination. Termination should be best effort,
if the instance is missing, that must mean it is terminated.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>